+++ src/Exceptions/InvalidDataException.php
@@ -4,7 +4,4 @@ namespace Luur\Amplitude\Exceptions;
 
 use Exception;
 
-class InvalidDataException extends Exception
-{
-
-}
+class InvalidDataException extends Exception {}